+++ DrGym/src/main/java/com/drgym/drgym/model/PostComment.java
@@ -5,25 +5,25 @@ import java.time.LocalDateTime;
 
 
 public class PostComment {
-    private Integer id;
+    private Long id;
     private String content;
     private String author;
     private LocalDateTime date;
 
     public PostComment() {}
 
-    public PostComment(Integer id, String content, String author, LocalDateTime date) {
+    public PostComment(Long id, String content, String author, LocalDateTime date) {
         this.id = id;
         this.content = content;
         this.author = author;
         this.date = date;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
